Fix the dings notification indicator
